refactor: use a single template for deployments
Previously, we used different templates for web vs background, which would eventually cause drift. Instead, we now use a single template, and remove the `port` stanza where it is not needed.

Note that this change also moved the templates to json, which required separating out the `service` template. This is good because we now _also_ track that the service has been rolled out correctly.